pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry]
name = "arms"
version = "v0.0.0"
description = ""
authors = ["Lawrence <lawrence@cinnamon.is>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12, <3.14"
aiohttp = "^3.10.10"
aiofiles = "^24.1.0"
pydantic = "^2.9.2"
pydantic-settings = "^2.6.1"

google-api-python-client = { version = "^2.151.0", optional = true }
google-auth-oauthlib = { version = "^1.2.1", optional = true }

[tool.poetry.extras]
google = ["google-api-python-client", "google-auth-oauthlib"]

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"

[tool.poetry.group.dev.dependencies]
mypy = "^1.13.0"
google-api-python-client-stubs = "^1.28.0"
types-beautifulsoup4 = "^4.12.0.20241020"

[tool.setuptools.package-data]
arms = ["py.typed"]

[tool.poetry-dynamic-versioning]
enable = true

[tool.ruff]
line-length = 79

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "ANN401",  # any-type
    "C901",  # complex-structure
    "D100",  # undocumented-public-module
    "D101",  # undocumented-public-class
    "D102",  # undocumented-public-method
    "D103",  # undocumented-public-function
    "D104",  # undocumented-public-package
    "D105",  # undocumented-magic-method
    "D106",  # undocumented-public-nested-class
    "D107",  # undocumented-public-init
    "D203",  # one-blank-line-before-class
    "D213",  # multi-line-summary-second-line
    "EM101",  # raw-string-in-exception
    "EM102",  # f-string-in-exception
    "FA102",  # future-required-type-annotation
    "FIX002",  # line-contains-todo
    "N815",  # mixed-case-variable-in-class-scope
    "PLR0911",  # too-many-return-statements
    "PLR0913",  # too-many-arguments
    "RUF001",  # ambiguous-unicode-character-string
    "TD002",  # missing-todo-author
    "TD003",  # missing-todo-link
    "TID252",  # relative-imports
    "TRY003",  # raise-vanilla-args
]

[tool.ruff.lint.per-file-ignores]
"tests/**.py" = [
  "S101",  # assert
]

[tool.mypy]
python_version = "3.11"
files = ["arms"]
check_untyped_defs = true
disallow_untyped_defs = true  # for strict mypy: (this is the tricky one :-))
disallow_untyped_calls = true
disallow_any_generics = true
ignore_missing_imports = true
no_implicit_reexport = true
warn_redundant_casts = true
warn_unused_ignores = true
plugins = ["pydantic.mypy"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.pdm.scripts]
lint = "ruff check --fix"
